+++ server/src/main/java/com/linbit/linstor/netcom/PeerTask.java
@@ -194,6 +194,28 @@ public class PeerTask implements Peer
     }
 
     @Override
+    public @Nullable ApiConsts.Platform getPlatform()
+    {
+        return null;
+    }
+
+    @Override
+    public void setPlatform(@Nullable ApiConsts.Platform platform)
+    {
+    }
+
+    @Override
+    public @Nullable String getOsVariant()
+    {
+        return null;
+    }
+
+    @Override
+    public void setOsVariant(@Nullable String osVariant)
+    {
+    }
+
+    @Override
     public boolean isConnected(boolean ensureAuthenticatedRef)
     {
         return false;